+++ Changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Form hooks to fix file upload plugin
+- fixed inability to use multiple upload plugins on one resource
 
 ## [v1.5.5]